User&priviledges mgmgt

+++ book_list.php
@@ -2,6 +2,7 @@
 <?php
 
     require_once 'include/user.php';
+    require_once 'user_required.php';
     
 
     $query = $db->prepare( 
@@ -29,11 +30,15 @@
                 <input class="form-control" id="searchBar" type="text" placeholder="Vyhledat knihu, autora, žánr">
             </div>
             <form action="users.php" method="GET" class="col text-right py-2 mr-3">
-            <a href="new_book.php" class="btn btn-success px-4 mt-1">Přidat knihu</a>
+            <?php
+            if($role > 1){
+            echo '<a href="new_book.php" class="btn btn-success px-4 mt-1">Přidat knihu</a>';
+            }
+            ?>
             </form>
         </div>
         
-        <div class="col w-100 px-0">
+    <div class="col w-100 px-0 pb-5">
         <ul class="list-group mx-3" id="searchList">
         <?
         foreach($book_list as $book){
@@ -54,8 +59,10 @@
             echo '          <div class="row text-right">';
             echo '              <div class="col text-right">';
             echo '                  <a href="./reservation.php?bookID='.$book['bookID'].'" class="btn btn-info btn-sm">Vypůjčit</a>';
+            if($role > 1){
             echo '                  <a href="./books_edit.php?bookID='.$book['bookID'].'" class="btn btn-light btn-sm">Upravit</a>';
             echo '                  <a href="./book_delete.php?bookID='.$book['bookID'].'" class="btn btn-danger btn-sm">Smazat</a>';
+            }
             echo '              </div>';
             echo '          </div>';
             echo '      </div>';
@@ -63,9 +70,9 @@
             echo '  </div>';
             echo '  </li>';
         }
-        echo '</div>';
         ?>
         </ul>
+    </div>
         
 <?php
     include 'include/footer.php';
